+++ src/intent_handling/cadocs_intents.py
@@ -1,0 +1,8 @@
+import enum
+
+
+class CadocsIntents(enum.Enum):
+   GetSmells = "get_smells"
+   GetSmellsDate = "get_smells_date"
+   Report = "report"
+   Info = "info"
