+++ tests/api_test/slack_api_connection_unit_test.py
@@ -1,6 +1,6 @@
 from src import slack_api_connection, service
 from src.service import utils
-from service.utils import CadocsIntents
+from intent_handling.cadocs_intents import CadocsIntents
 from flask import Flask, make_response
 from slack_api_connection import app, Cadocs
 import json
