+++ tests/intent_test/IntentResolver_unit_test.py
@@ -1,8 +1,9 @@
-from src import intent_resolver, service, tools, tool_selector
-from intent_resolver import IntentResolver
-from service.utils import CadocsIntents
-from tool_selector import ToolSelector
-from tools import CsDetectorTool
+from intent_handling import intent_resolver, tool_selector, tools
+from src import service
+from intent_handling.intent_resolver import IntentResolver
+from intent_handling.cadocs_intents import CadocsIntents
+from intent_handling.tool_selector import ToolSelector
+from intent_handling.tools import CsDetectorTool
 from unittest.mock import Mock
 import pytest
 
@@ -47,7 +48,7 @@ class TestIntentResolverUT:
 
         # Mock the build_message method of the cadocs_messages module
         mocker.patch(
-            'intent_resolver.cadocs_messages.build_cs_message', return_value="Test OK")
+            'intent_handling.intent_resolver.cadocs_messages.build_cs_message', return_value="Test OK")
         result = intent_resolver_instance.build_message(
             "Test OK", user, "channel", CadocsIntents.GetSmells, ["https://github.com/tensorflow/ranking"])
 
@@ -62,7 +63,7 @@ class TestIntentResolverUT:
 
         # Mock the build_message method of the cadocs_messages module
         mocker.patch(
-            'intent_resolver.cadocs_messages.build_cs_message', return_value="Test OK")
+            'intent_handling.intent_resolver.cadocs_messages.build_cs_message', return_value="Test OK")
         result = intent_resolver_instance.build_message(
             "Test OK", user, "channel", CadocsIntents.GetSmellsDate, ["https://github.com/tensorflow/rankin, 12/12/2022"])
 
@@ -77,7 +78,7 @@ class TestIntentResolverUT:
 
         # Mock the build_message method of the cadocs_messages module
         mocker.patch(
-            'intent_resolver.cadocs_messages.build_report_message', return_value="Test OK")
+            'intent_handling.intent_resolver.cadocs_messages.build_report_message', return_value="Test OK")
         result = intent_resolver_instance.build_message(
             "Test OK", user, "channel", CadocsIntents.Report, ["https://github.com/tensorflow/ranking", "12/12/2022", "report"])
 
@@ -92,7 +93,7 @@ class TestIntentResolverUT:
 
         # Mock the build_message method of the cadocs_messages module
         mocker.patch(
-            'intent_resolver.cadocs_messages.build_info_message', return_value="Test OK")
+            'intent_handling.intent_resolver.cadocs_messages.build_info_message', return_value="Test OK")
         result = intent_resolver_instance.build_message(
             "Test OK", user, "channel", CadocsIntents.Info, [])
